Title: 배열의 특정 위치 찾기, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 241022/배열의 특정 위치 찾기/find-specific-location-fo-array.cpp
@@ -24,6 +24,6 @@ int main() {
             cnt++;
         }
     }
-    printf ("%d %.1f", evensum, (float)(oddsum / cnt));
+    printf ("%d %.1f", evensum, (double)(oddsum / cnt));
     return 0;
 }
